v2.2.2 — custom theme builder, BotBooru, inline images, stronger no-speak
## Custom Theme Builder
- New "Create Theme" screen with HSV color wheel pickers for all colors
- Live split preview: main menu example + avatar shape chips + chat example
- Background image support for main menu (fill/fit/stretch, opacity slider)
- Avatar shape selector: Circle, Rounded Square, Square (removed Name Only)
- Save theme as ST-compatible JSON; accessible from Theme Settings screen

## BotBooru Browser
- Added BotBooru as third option in "Search cards" dropdown (alongside CharaVault and RisuRealm)
- Embedded WebView browser at botbooru.com
- Blob URL download fix: hooks URL.createObjectURL and HTMLAnchorElement.prototype.click
  before page scripts run so Download PNG is intercepted before revokeObjectURL fires

## Inline Image Rendering in Chat
- Markdown images ![alt](url) rendered as inline images inside chat bubbles
- HTML <img src="..."> tags (including malformed variants) also rendered inline
- Streaming bubble strips image syntax from the text preview

## Stronger "Don't Speak for User" Constraint
- Injected at both system-prompt level and post-history level
- Stronger wording with explicit NEVER BREAK THIS RULE framing

## Image Generation
- setModel() implemented for SD WebUI / Forge backend

+++ app/src/main/kotlin/com/pockettavern/app/data/remote/dto/forge/ForgeDtos.kt
@@ -85,7 +85,7 @@ data class OptionsDto(
     @SerialName("sd_vae")
     val sdVae: String? = null,
     @SerialName("CLIP_stop_at_last_layers")
-    val clipStopAtLastLayers: Int? = null
+    val clipStopAtLastLayers: Double? = null
 )
 
 @Serializable
